Added source tab menu, moved tools check
The source tab popup menu allows closing a file in a taba and opening a
file in the same or a new tab.

The tools check in mainwindow.cpp was after starting gdb.  On a computer
with no gdb, ebe would get stuck consuming CPU time with no message.
I moved it to before starting gdb.