Changed uLong/sLong types; ebedecl fix
It is clear that the long long types are 64 bits on all relevent
systems and there is no need for #if's related to these types.
This eliminates printf warnings on some systems.

Ebedecl used %precedence in parse.y which was not a bison command
in RHEL.  %left works just as well and is available in RHEL bison.